requirements.txt
Django==4.1
djangorestframework==3.13.1
drf-spectacular==0.23.1
pillow==9.2.0
django-cors-headers==3.13.0
django-ckeditor == 6.5.1
django-money==3.0.0
django_jalali==6.0.0
kavenegar==1.1.2
django-taggit==3.1.0
flower
django-environ==0.8.1
redis==4.5.1
celery==5.2.7
django-celery-beat==2.4.0
django-extensions
django-jalali==6.0.0
django-jalali-date==1.0.2
psycopg2==2.9.3
model-bakery==1.11.0
black==23.1.0
django-filter